GEODE-9092: Resolve package splitting geode-common
