+++ geode-common/src/main/java/org/apache/geode/internal/utils/Retry.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal;
+package org.apache.geode.internal.utils;
 
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
